+++ R/as_task_cmprsk.R
@@ -27,13 +27,21 @@ as_task_cmprsk.TaskCompRisks = function(x, clone = FALSE, ...) {
 #'   Id for the new task.
 #'   Defaults to the (deparsed and substituted) name of `x`.
 #' @export
-as_task_cmprsk.data.frame = function(x, time = "time", event = "event",
-                                      id = deparse(substitute(x)), ...) {
+as_task_cmprsk.data.frame = function(
+  x,
+  time = "time",
+  event = "event",
+  id = deparse(substitute(x)),
+  ...
+) {
   assert_data_frame(x, min.rows = 1L, min.cols = 1L, col.names = "unique")
 
   ii = which(map_lgl(keep(x, is.double), anyInfinite))
   if (length(ii)) {
-    warningf("Detected columns with unsupported Inf values in data: %s", str_collapse(names(ii)))
+    warning_input(
+      "Detected columns with unsupported Inf values in data: %s",
+      str_collapse(names(ii))
+    )
   }
 
   TaskCompRisks$new(id = id, backend = x, time = time, event = event, ...)
@@ -41,7 +49,12 @@ as_task_cmprsk.data.frame = function(x, time = "time", event = "event",
 
 #' @rdname as_task_cmprsk
 #' @export
-as_task_cmprsk.DataBackend = function(x, time = "time", event = "event",
-                                       id = deparse(substitute(x)), ...) {
+as_task_cmprsk.DataBackend = function(
+  x,
+  time = "time",
+  event = "event",
+  id = deparse(substitute(x)),
+  ...
+) {
   TaskCompRisks$new(id = id, backend = x, time = time, event = event, ...)
 }